move maybe_report_invalid_custom_discriminants to feature_gate
